feat(appkit): infer numeric SQL type for sql.number(), add typed variants
Currently sql.number() unconditionally produces __sql_type: "NUMERIC",
which Databricks SQL binds as DECIMAL(10,0). That breaks LIMIT and
OFFSET — Spark requires an integer-typed expression and rejects the
parameter with INVALID_LIMIT_LIKE_EXPRESSION.DATA_TYPE — and silently
truncates non-integer JS numbers (sql.number(3.14) sent value="3.14"
into a DECIMAL(10,0) slot).

Two changes:

1. sql.number() now infers the wire type from the value:
   - integer JS number -> BIGINT
   - non-integer JS number -> DOUBLE
   - numeric string -> NUMERIC (preserves caller's precision intent;
     a string is an explicit choice to skip JS-number coercion)

   Picks the most-precise type the value can losslessly represent.
   Spark coerces numeric types implicitly so existing queries against
   BIGINT/DECIMAL/DOUBLE columns keep working, plus LIMIT now accepts
   sql.number(10) directly.

2. Typed variants for callers who need to override the inference:
   - sql.int(value)     -> INT
   - sql.bigint(value)  -> BIGINT (also accepts JS bigint for values
                          beyond Number.MAX_SAFE_INTEGER)
   - sql.float(value)   -> FLOAT
   - sql.double(value)  -> DOUBLE
   - sql.decimal(value) -> NUMERIC (decimal precision preserved)

   sql.int() and sql.bigint() reject non-integer inputs at the
   helper boundary so the failure surfaces early, before the wire.

SQLNumberMarker.__sql_type widens from "NUMERIC" to a union of the
numeric SQL types. Non-breaking for callers: any code that previously
held an SQLNumberMarker still type-checks (the union is wider in
return position). The two unit tests that pinned `type: "NUMERIC"`
for sql.number(integer) are updated to expect BIGINT.

Discovered while building a parameterized analytics query against
LIMIT — the bare `LIMIT :limit` case is the most visible failure but
the underlying issue affects any query where the column type matters.

+++ packages/appkit/src/plugins/analytics/tests/query.test.ts
@@ -32,7 +32,7 @@ describe("QueryProcessor", () => {
       expect(result.statement).toBe(query);
       expect(result.parameters).toHaveLength(2);
       expect(result.parameters).toEqual([
-        { name: "user_id", value: "123", type: "NUMERIC" },
+        { name: "user_id", value: "123", type: "BIGINT" },
         { name: "name", value: "Alice", type: "STRING" },
       ]);
     });
@@ -229,7 +229,7 @@ describe("QueryProcessor", () => {
       expect(result.parameters[0]).toEqual({
         name: "age",
         value: "25",
-        type: "NUMERIC",
+        type: "BIGINT",
       });
     });
